% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{apply_skip_logic}
\alias{apply_skip_logic}
\title{Apply skip logic}
\usage{
apply_skip_logic(data, config)
}
\arguments{
\item{data}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Applies skip logic to dataset, returning NA for downstream questions to adhere to survey streaming.
}
