% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{axis_settings}
\alias{axis_settings}
\title{Axis settings}
\usage{
axis_settings(xlab, ylab, font_size)
}
\arguments{
\item{xlab}{see @plot_freqs}

\item{ylab}{see @plot_freqs}

\item{font_size}{see @plot_freqs}
}
\value{
list of axis settings.
}
\description{
Standard X (categorical) and Y (percentage scale) axis settings for use with the plotting functions.
}
