% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calculate_bases}
\alias{calculate_bases}
\title{Calculate bases for Likert chart}
\usage{
calculate_bases(data, mid, neutral_mid)
}
\arguments{
\item{data}{see @plot_likert}

\item{mid}{see @plot_likert}

\item{neutral_mid}{see @plot_likert}
}
\value{
bar starting locations
}
\description{
Calculates starting locations for bars on the Likert chart.
}
