% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARS-package.R
\docType{package}
\name{CARS-package}
\alias{CARS}
\alias{CARS-package}
\title{CARS: Coding in Analysis and Research Survey (CARS)}
\description{
Analysis code for the coding in analysis and research survey (CARS).
}
\author{
\strong{Maintainer}: Analysis standards and pipelines \email{asap@ons.gov.uk}

}
\keyword{internal}
