% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARS-package.R
\docType{package}
\name{CARS}
\alias{CARS}
\title{CARS}
\description{
Analysis for the Coding in Analysis and Research Survey (CARS) wave 4 (2022).
The package produces frequencies and outputs for CARS data.
}
