% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-routing.R
\name{check_skip_logic}
\alias{check_skip_logic}
\title{Check skip logic}
\usage{
check_skip_logic(data, condition, skipped_cols)
}
\arguments{
\item{data}{data.frame}

\item{condition}{logical vector. Example: data$row == "skip response"}

\item{skipped_cols}{character. questions that should have been skipped if condition != TRUE}
}
\value{
list of rows failing the check
}
\description{
Checks whether the skip logic was followed correctly. Backtracking while filling the survey can result in inconsistent response sets.
This check returns row numbers where questions which should have been skipped contain anything other than NA.
}
