% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
Recategorise department, workplace and first_learned data
}
