% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_departments}
\alias{clean_departments}
\title{Clean department data}
\usage{
clean_departments(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
add NHS to department list and merge departments where needed.
}
