% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_first_learned}
\alias{clean_first_learned}
\title{Clean first_learned data}
\usage{
clean_first_learned(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
categorise 'other' responses.
}
