% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_quality_qs}
\alias{clean_quality_qs}
\title{Clean quality question columns to remove whitespace}
\usage{
clean_quality_qs(data)
}
\arguments{
\item{data}{CARS 2024 dataset}
}
\value{
cleaned quality question columns
}
\description{
Clean quality question columns to remove whitespace
}
