% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_workplace}
\alias{clean_workplace}
\title{Clean workplace data}
\usage{
clean_workplace(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
reclassify 'other' text responses into CS/NHS
}
