% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{compare_models}
\alias{compare_models}
\title{Compare the AIC scores of several models}
\usage{
compare_models(models)
}
\arguments{
\item{models}{a vector of fitted models}
}
\value{
data.frame object of model's AIC scores in ascending order
}
\description{
Compare the AIC scores of several models
}
