% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{count_to_prop}
\alias{count_to_prop}
\title{Convert frequencies to proportions}
\usage{
count_to_prop(data)
}
\arguments{
\item{data}{frequency table with three columns (can be of any name): name, value and count}
}
\value{
input data with the third column as proportion (0-1)
}
\description{
Convert frequencies to proportions
}
