% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{create_filtered_pages}
\alias{create_filtered_pages}
\title{create filtered pages}
\usage{
create_filtered_pages(
  data,
  type = c("professions", "departments"),
  qmd_path = "quarto/main",
  template_path = "quarto/templates"
)
}
\arguments{
\item{data}{full, cleaned CARS dataset. Only needed for department pages.}

\item{type}{type of filtered page - department or profession}

\item{qmd_path}{qmd site folder (without trailing "/")}

\item{template_path}{template folder (without trailing "/")}
}
\description{
Creates summary statistics pages for professions/departments. Also creates index pages.
}
