% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{create_prof_pages}
\alias{create_prof_pages}
\title{Create profession page}
\usage{
create_prof_pages(
  template_path = "quarto/templates/summary.qmd",
  output_path = "quarto/main/professions"
)
}
\arguments{
\item{template_path}{location of the quarto template}

\item{output_path}{output folder. Do not end with "\". Proceed with caution:
existing folders will be overwritten, including any files}
}
\description{
Create profession breakdown pages from template
}
