% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidy_cross_table.R
\name{create_tidy_cross_table}
\alias{create_tidy_cross_table}
\title{Create tidy cross table}
\usage{
create_tidy_cross_table(data, col1, col2, levels1, levels2)
}
\arguments{
\item{data}{pre-processed CARS data set}

\item{col1}{first column of interest}

\item{col2}{column to cross-tabulate first column against}

\item{levels1}{factor levels for col1}

\item{levels2}{factor levels for col2}
}
\value{
data.frame
}
\description{
Returns a cross table in tidy data format.
}
