% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidy_freq_table.R
\name{create_tidy_freq_table}
\alias{create_tidy_freq_table}
\title{Create tidy frequency table}
\usage{
create_tidy_freq_table(data, questions, levels, labels)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after pre-processing}

\item{questions}{columns to filter data on}

\item{levels}{all possible factor values in the filtered columns}

\item{labels}{labels to rename the column headers}
}
\value{
data.frame
}
\description{
Returns a frequency table in tidy data format.
}
