% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars.R
\name{derive_rap_champ_status}
\alias{derive_rap_champ_status}
\title{Derive RAP Champion status}
\usage{
derive_rap_champ_status(data)
}
\arguments{
\item{data}{a date frame containing cleaned CARS wave 5 data}
}
\value{
dataframe containing the additional RAP Champion status columns
}
\description{
Derive RAP Champion status column from existing variables and add to the dataframe.
}
