% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tables.R
\name{df_to_crosstab}
\alias{df_to_crosstab}
\title{Convert data frame to crosstab}
\usage{
df_to_crosstab(data)
}
\arguments{
\item{data}{tidy frequency table}
}
\value{
wide data.frame
}
\description{
Converts tidy data with two factor columns to a cross tabulation (wide data).
}
