% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{display_prof_langs}
\alias{display_prof_langs}
\title{Display programming languages filtered by profession}
\usage{
display_prof_langs(table, prof, prof_col)
}
\arguments{
\item{table}{frequency table (language capability, see frequency table functions)}

\item{prof}{profession name}

\item{prof_col}{profession column name}
}
\value{
HTML output - table and chart
}
\description{
Display programming languages filtered by profession
}
