% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARS.R
\name{epsest.func}
\alias{epsest.func}
\title{Estimation of the non-null proportion}
\usage{
epsest.func(x, u, sigma)
}
\arguments{
\item{x}{the corresponding vector to be estimated}

\item{u}{the mean of the null distribution}

\item{sigma}{the standard deviation of the null distribution}
}
\value{
a value indicating the estimated non-null proportion
}
\description{
Estimates the proportion of non-nulls.
}
\examples{
X <- rep(c(0,2),c(800,200))+rnorm(1000);
epsest.func(X,0,1);

}
