% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sizes.R
\name{expected_sample_sizes}
\alias{expected_sample_sizes}
\title{Sample sizes for table/plot outputs}
\usage{
expected_sample_sizes(data)
}
\arguments{
\item{data}{full CARS dataset after cleaning}
}
\value{
list of sample sizes
}
\description{
Sample sizes for table/plot outputs
}
