% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scales.R
\name{get_2colour_scale}
\alias{get_2colour_scale}
\title{Create 2 colour scale palette}
\usage{
get_2colour_scale(
  n,
  colour1 = c(18, 67, 109),
  colour2 = c(244, 106, 37),
  colour_names = TRUE
)
}
\arguments{
\item{n}{the number of colours needed}

\item{colour1}{the first colour in the scale - a numeric vector representing red, green and blue (max 255)}

\item{colour2}{the last colour in the scale - numeric vector representing red, green and blue (max 255).}

\item{colour_names}{whether to return HTML colour names. TRUE by default.}
}
\value{
vector of hexadecimal colours containing shades between the two selected colours
}
\description{
Creates a list of  colour names (RGB). Generates a scale between two colours. By default these are the
shades of orange and blue used in the analysis function colour scheme.
}
