% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{get_ci}
\alias{get_ci}
\title{Get percentage and confidence interval}
\usage{
get_ci(table, freq_col, n_col)
}
\arguments{
\item{table}{frequency table with sample column}

\item{freq_col}{frequency column index}

\item{n_col}{sample column index}
}
\value{
data frame with new percentage and upper/lower ci columns
}
\description{
Get percentage and confidence interval
}
