% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scales.R
\name{get_gradient}
\alias{get_gradient}
\title{Create single colour gradient}
\usage{
get_gradient(n, colour = c(18, 67, 109), colour_names = TRUE)
}
\arguments{
\item{n}{the number of colours needed}

\item{colour}{the first colour of the gradient - numeric vector with three values representing red, blue and green (max 255)}

\item{colour_names}{whether to return HTML colour names. TRUE by default.}
}
\value{
RGB colours
}
\description{
Creates a list of colours for plotting. The list contains n shades of the original colour, increasing in brightness.
}
