% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{get_question_data}
\alias{get_question_data}
\title{Get question metadata}
\usage{
get_question_data(config, question)
}
\arguments{
\item{config}{CARS config file}

\item{question}{question name as written in config file}
}
\value{
List of data cols, labels (if multiple cols), and factor levels for the specified question
}
\description{
Get question metadata
}
