% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{get_tidy_data_api}
\alias{get_tidy_data_api}
\title{Get tidy data from API}
\usage{
get_tidy_data_api(...)
}
\arguments{
\item{...}{optional: arguments to be passed to \link{ingest}}
}
\value{
tidied CARS dataset
}
\description{
API function for data ingest. Loads data from the smartsurvey API and convert to a tidy data frame.
}
