% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{get_tidy_data_file}
\alias{get_tidy_data_file}
\title{Get tidy data from file}
\usage{
get_tidy_data_file(...)
}
\arguments{
\item{...}{optional: arguments to be passed to \link{ingest}}
}
\value{
tidied CARS dataset
}
\description{
API function for data ingest. Loads data from csv and returns pre-processed data frame.
}
