% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest}
\alias{ingest}
\title{Ingest smartsurvey data}
\usage{
ingest(
  survey = "1376897",
  token = Sys.getenv("CARS_TOKEN"),
  secret = Sys.getenv("CARS_SECRET"),
  proxies = Sys.getenv("alt_proxy"),
  export
)
}
\arguments{
\item{survey}{the survey ID (character string/numeric). Defaults to "1167489".}

\item{token}{the API token (character string). Loaded from environment variable by default.}

\item{secret}{the secret API token (character string). Loaded from environment variable by default.}

\item{proxies}{proxy addresses (string). Loads from the user environment by default. Expects a string in the format "ip.address:port; alt.ip.address:port".}

\item{export}{the export ID (character string/numeric).}
}
\value{
the exported data as a dataframe
}
\description{
Download smartsurvey export via the API. Download the exported data from smartsurvey using the API. Use convert_raw() to convert the API response to a data.frame.
Note: the first API request in a session will typically fail.
}
