% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARS.R
\name{lin.itp}
\alias{lin.itp}
\title{Linear interpolation}
\usage{
lin.itp(x, X, Y)
}
\arguments{
\item{x}{the coordinates of points where the density needs to be interpolated}

\item{X}{the coordinates of the estimated densities}

\item{Y}{the values of the estimated densities}
}
\value{
the interpolated densities
}
\description{
Interpolates desired vector given density estimation.
}
\examples{
X <- seq(-10,10,length.out=20);
Y <- dnorm(X);
x <- seq(-10,10,length.out=100);
lin.itp(x,X,Y)

}
