% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_freqs}
\alias{plot_freqs}
\title{Plot frequency graph}
\usage{
plot_freqs(
  data,
  config,
  question,
  colour,
  break_q_names_col,
  type = c("bar", "line"),
  max_lines = 2,
  xlab = "",
  ylab = "",
  font_size = 12,
  orientation = c("v", "h"),
  ...
)
}
\arguments{
\item{data}{Frequency data with sample column (data frame). Expected input: data.frame(categories = c(), frequencies = c(), count = c(), sample = c())}

\item{colour}{Colour name. Defaults to blue (see @get_gradient())}

\item{break_q_names_col}{applies break_q_names to the column. Not applied by default}

\item{type}{optional: chart type ("bar" or "line").}

\item{max_lines}{maximum number of lines. Int, defaults to 2/ See carsurvey::break_q_names()}

\item{xlab}{X axis title (optional)}

\item{ylab}{Y axis title (optional)}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce bar chart (plotly) for single factor frequency data.
}
