% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_grouped}
\alias{plot_grouped}
\title{Plot grouped frequency graph}
\usage{
plot_grouped(
  data,
  n,
  break_q_names_col,
  max_lines = 2,
  xlab = "",
  ylab = "",
  font_size = 12,
  orientation = c("v", "h"),
  ...
)
}
\arguments{
\item{data}{Frequency data (data frame). 3 columns - category names, groups and frequencies.}

\item{n}{sample size}

\item{break_q_names_col}{applies break_q_names to the column. Not applied by default}

\item{max_lines}{maximum number of lines. Int, defaults to 2/ See carsurvey::break_q_names()}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default.}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce bar chart (plotly) for frequency data with grouping variable.
}
