% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_likert}
\alias{plot_likert}
\title{Plot Likert graph}
\usage{
plot_likert(
  data,
  mid,
  n,
  break_q_names_col,
  max_lines = 2,
  xlab = "",
  ylab = "",
  font_size = 12,
  neutral_mid = TRUE,
  ...
)
}
\arguments{
\item{data}{Frequency data for likert quesitons (data frame). 3 columns: question, answer option, frequency.}

\item{mid}{the mid-point of the scale. should be higher than 2 and lower than the number of answers.}

\item{n}{sample size}

\item{break_q_names_col}{applies break_q_names to the column. Not applied by default}

\item{max_lines}{changes maximum lines text can go over}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{font_size}{minimum font size for the plot (numeric).}

\item{neutral_mid}{whether the middle of the scale should be a neutral category (logical). TRUE by default}

\item{...}{additional plot_ly arguments}
}
\value{
bar chart
}
\description{
Produce Likert stacked bar chart (plotly). At least 2 questions per plot.
}
