% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stacked}
\alias{plot_stacked}
\title{Plot stacked bar graph}
\usage{
plot_stacked(
  data,
  n,
  break_q_names_col,
  type = c("bar", "line"),
  max_lines = 2,
  xlab = "",
  ylab = "",
  colour_scale = c("2gradients", "gradient", "scale", "3scale"),
  font_size = 12,
  neutral_mid = TRUE,
  orientation = c("h", "v"),
  ...
)
}
\arguments{
\item{data}{Frequency data for stacked bar chart (data frame). 3 columns: variable 1, variable 2 and values (tidy data)}

\item{n}{sample size}

\item{break_q_names_col}{applies break_q_names to the column. Not applied by default}

\item{type}{optional: chart type ("bar" or "line").}

\item{max_lines}{maximum number of lines. Int, defaults to 2/ See carsurvey::break_q_names()}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{colour_scale}{type of colour scale ("gradient", "scale" or "2gradients"). See get_gradient(), get_2colour_scale() and get_2colour_gradients().}

\item{font_size}{minimum font size for the plot (numeric).}

\item{neutral_mid}{whether the midpoint of the colour scale should be neutral ("2gradients" scale only). TRUE by default}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce stacked bar chart (plotly).
}
