% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{prop_by_group}
\alias{prop_by_group}
\title{Convert frequencies to proportions}
\usage{
prop_by_group(data)
}
\arguments{
\item{data}{frequency table with three columns (can be of any name): name, value and count}
}
\value{
input data with the third column as proportion (0-1)
}
\description{
Convert frequencies to proportions
}
