% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{read_file}
\alias{read_file}
\title{Read file}
\usage{
read_file(name, dir = Sys.getenv("CARS_DATA_DIR"))
}
\arguments{
\item{name}{file name}

\item{dir}{data folder (defaults to CARS_DATA_DIR environment vairable)}
}
\value{
unprocessed CARS data
}
\description{
Wrapper for read.csv - loads data using path saved in environment variable.
}
