% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_site_yml.R
\name{read_site_yml}
\alias{read_site_yml}
\title{Read RMarkdown site yaml}
\usage{
read_site_yml(filename)
}
\arguments{
\item{filename}{the yaml file name (including path)}
}
\value{
list of arguments
}
\description{
Read the rmarkdown site yaml and return list of navbar arguments. Used to render the custom navigation bar. The yaml file must follow the same syntax expected by RMarkdown/quarto.
}
