% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_prof_page}
\alias{render_prof_page}
\title{render profession page}
\usage{
render_prof_page(data, prof, title, output_path, path = "quarto/template")
}
\arguments{
\item{data}{full pre-processed CARS dataset}

\item{prof}{profession to include in the report}

\item{title}{page title}

\item{output_path}{output path (will overwrite existing outputs)}

\item{path}{quarto input}
}
\description{
render CARS publication from quarto
}
