% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_site}
\alias{render_site}
\title{render site}
\usage{
render_site(data, path = "quarto/main", output_path = "docs/")
}
\arguments{
\item{data}{full pre-processed CARS dataset}

\item{path}{quarto input}

\item{output_path}{output path (will overwrite existing outputs). Should match the path set in the quarto site yaml}
}
\description{
render CARS publication from quarto
}
