% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{replace_apostrophes}
\alias{replace_apostrophes}
\title{Replace apostrophes}
\usage{
replace_apostrophes(data)
}
\arguments{
\item{data}{tidy CARS dataset}
}
\value{
data.frame
}
\description{
Replaces @SQ@ in the data with '.
}
