% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{sample_sizes}
\alias{sample_sizes}
\title{Sample sizes for table/plot outputs}
\usage{
sample_sizes(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
list of sample sizes
}
\description{
Sample sizes for table/plot outputs
}
