% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_axis_range}
\alias{set_axis_range}
\title{Set axis range}
\usage{
set_axis_range(plot, min, max, axis = c("x", "y"))
}
\arguments{
\item{plot}{plotly object}

\item{axis}{optional: defaults to "x"}
}
\value{
list of parameters for plotly annotation
}
\description{
Sets x or y axis range for plotly objects
}
