% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_error_bars}
\alias{set_error_bars}
\title{Set up error bars}
\usage{
set_error_bars(lower_ci, upper_ci)
}
\arguments{
\item{lower_ci}{lower confidence interval}

\item{upper_ci}{upper confidence interval}
}
\value{
list of parameters for plotly error bars
}
\description{
Formats plotly error bar settings. Can be used with plots
}
