% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_ability_change}
\alias{summarise_ability_change}
\title{Summarise ability change frequency}
\usage{
summarise_ability_change(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for ability change
}
