% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_access_git}
\alias{summarise_access_git}
\title{Summarise access to git}
\usage{
summarise_access_git(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone has access to git
}
