% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_adv_score_by_understanding}
\alias{summarise_adv_score_by_understanding}
\title{Compare advanced RAP score to understanding of key RAP components}
\usage{
summarise_adv_score_by_understanding(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for advanced rap score compared with understanding of key RAP components
}
