% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_basic_score_by_imp}
\alias{summarise_basic_score_by_imp}
\title{Compare basic RAP score to implementation of RAP}
\usage{
summarise_basic_score_by_imp(data)
}
\arguments{
\item{data}{full CARS datasetafter pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for basic rap score compared with implementation of RAP
}
