% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_basic_score_by_understanding}
\alias{summarise_basic_score_by_understanding}
\title{Compare basic RAP score to understanding of key RAP components}
\usage{
summarise_basic_score_by_understanding(data)
}
\arguments{
\item{data}{full CARS datasetafter pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for basic rap score compared with understanding of key RAP components
}
