% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_cap_change_by_CS_grade}
\alias{summarise_cap_change_by_CS_grade}
\title{Summarise capability change by CS grade}
\usage{
summarise_cap_change_by_CS_grade(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate the cross tab of capability change by CS grade
}
