% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_cap_change_by_grade}
\alias{summarise_cap_change_by_grade}
\title{Summarise capability change by CS grade}
\usage{
summarise_cap_change_by_grade(
  data,
  config,
  question1,
  question2,
  prop = TRUE,
  sample = FALSE
)
}
\arguments{
\item{data}{cleaned CARS dataset}

\item{config}{CARS config}

\item{question1}{ability change}

\item{question2}{cs_grade}

\item{prop}{whether to return proportion data (0-1). TRUE by default. Assumes mutually exclusive response options.}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate the cross tab of capability change by CS grade
}
