% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_cap_change_by_line_manage}
\alias{summarise_cap_change_by_line_manage}
\title{Summarise capability change by management responsibility}
\usage{
summarise_cap_change_by_line_manage(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate the cross tab of capability change by management responsibility
}
