% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_ci}
\alias{summarise_ci}
\title{Summarise continuous integration frequency}
\usage{
summarise_ci(data)
}
\arguments{
\item{data}{full CARS datasetafter pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for continuous integration
}
