% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_code_freq}
\alias{summarise_code_freq}
\title{Summarise coding frequency}
\usage{
summarise_code_freq(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for coding frequency.
}
