% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_coding_practices}
\alias{summarise_coding_practices}
\title{Summarise data practices questions}
\usage{
summarise_coding_practices(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for data practices
}
