% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_dep_man}
\alias{summarise_dep_man}
\title{Summarise dependency management frequency}
\usage{
summarise_dep_man(data)
}
\arguments{
\item{data}{full CARS datasetafter pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for dependency management.
}
