% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_git}
\alias{summarise_git}
\title{Summarise access to git and knowledge of git}
\usage{
summarise_git(data, config, question, prop = TRUE, sample = TRUE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{config}{CARS config}

\item{question}{coding tools knowledge or access question, taken from config}

\item{prop}{returns proportion, TRUE by default}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone has access to git and calculate frequency table for if someone knows how to version control using git
}
