% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_heard_of_RAP_by_prof}
\alias{summarise_heard_of_RAP_by_prof}
\title{Summarise heard of RAP by profession}
\usage{
summarise_heard_of_RAP_by_prof(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of basic and advanced RAP score components
}
