% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_knowledge_git}
\alias{summarise_knowledge_git}
\title{Summarise knowledge of git}
\usage{
summarise_knowledge_git(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone knows how to version control using git
}
