% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_line_manage}
\alias{summarise_line_manage}
\title{Summarise manage someone who codes}
\usage{
summarise_line_manage(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone line manages someone who codes
}
