% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_multi_col_data}
\alias{summarise_multi_col_data}
\title{Generic summarise multi-column data into frequency table}
\usage{
summarise_multi_col_data(data, config, question, prop = TRUE, sample = FALSE)
}
\arguments{
\item{data}{cleaned CARS dataset}

\item{config}{CARS config}

\item{question}{question name taken from config}

\item{prop}{additionally returns value as a proportion. TRUE by default}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for multi-column data where no additional operations are needed
}
