% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_open_source_by_prof}
\alias{summarise_open_source_by_prof}
\title{Summarise open source practice by profession}
\usage{
summarise_open_source_by_prof(data)
}
\arguments{
\item{data}{CARS data (pre-processed)}
}
\value{
data.frame
}
\description{
only used the main summary page. Needs to be turned into wide data for html table.
}
