% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_os_vs_prop}
\alias{summarise_os_vs_prop}
\title{Summarise open source vs proprietary capability}
\usage{
summarise_os_vs_prop(data)
}
\arguments{
\item{data}{Full CARS dataset including previous waves}
}
\value{
data.frame
}
\description{
Calculate proportion of respondents who have capability in R/Python vs SAS/SPSS/stata
}
