% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_rap_awareness_over_time}
\alias{summarise_rap_awareness_over_time}
\title{RAP awareness over time}
\usage{
summarise_rap_awareness_over_time(data)
}
\arguments{
\item{data}{all wave data}
}
\value{
data frame
}
\description{
RAP awareness over time
}
