% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_rap_comp}
\alias{summarise_rap_comp}
\title{RAP score components}
\usage{
summarise_rap_comp(data, config, question, sample = TRUE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{config}{CARS config}

\item{question}{rap components question, taken from config}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of basic and advanced RAP score components
}
