% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_rap_knowledge}
\alias{summarise_rap_knowledge}
\title{Knowledge of RAP}
\usage{
summarise_rap_knowledge(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
Create a frequency table of knowledge of RAP
}
