% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_strategy_knowledge}
\alias{summarise_strategy_knowledge}
\title{Summarise Analysis Function RAP strategy knowledge}
\usage{
summarise_strategy_knowledge(data)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone heard of or read the RAP strategy
}
