% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_where_learned_code}
\alias{summarise_where_learned_code}
\title{Summarise where respondents learned to code}
\usage{
summarise_where_learned_code(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table of where respondents learned to code
}
