% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cars.R
\name{theme_cars}
\alias{theme_cars}
\title{A Cars Theme to ggplot2 Functions}
\usage{
theme_cars(transparent_background = T)
}
\arguments{
\item{transparent_background}{A boolean for whether or not to include
a transparent background. Set to false if sharing a .png file}
}
\value{
a gglot theme
}
\description{
This function provides a wrapper around some ggplot2 theme
arguments and creates a unified CARS graphing theme, along with
other color functions provided in caRs
}
