% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{tidy_colnames}
\alias{tidy_colnames}
\title{Tidy column names}
\usage{
tidy_colnames(raw_data)
}
\arguments{
\item{raw_data}{data frame returned by convert_raw()}
}
\value{
data frame with empty rows removed and tidied column names
}
\description{
Takes smartsurvey data as an input and converts to a tidier format, with question names removed from value rows.
}
