% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_upper_first}
\alias{to_upper_first}
\title{Capitalise the first character of a string}
\usage{
to_upper_first(text)
}
\arguments{
\item{text}{a character object of any length}
}
\value{
character object, with the first character of each string capitalised
}
\description{
Capitalise the first character of a string
}
