% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_2_preprocessing.R
\name{w2_enforce_streaming}
\alias{w2_enforce_streaming}
\title{Enforce streaming rules}
\usage{
w2_enforce_streaming(data)
}
\arguments{
\item{data}{the data which is the output from carsurvey2::tidy_ingest(data)}
}
\value{
the exported data after streaming rules as a dataframe
}
\description{
Input the data as a data frame and apply each of the streaming rules in turn.
}
