% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_2_preprocessing.R
\name{w2_rename_cols}
\alias{w2_rename_cols}
\title{Rename columns (wave 3)}
\usage{
w2_rename_cols(data)
}
\arguments{
\item{data}{CARS wave 3 (2021) survey data (data.frame).}
}
\value{
data.frame
}
\description{
add meaningful column names to dataset ingested from smartsurvey API.
}
