% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_codefreq_dep_man}
\alias{w3_enforce_codefreq_dep_man}
\title{Enforce code frequency streaming on dependency management questions (wave 3)}
\usage{
w3_enforce_codefreq_dep_man(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then dependency management questions is skipped
}
