% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_codefreq_rep_wf}
\alias{w3_enforce_codefreq_rep_wf}
\title{Enforce code frequency streaming on reproducible workflow questions (wave 3)}
\usage{
w3_enforce_codefreq_rep_wf(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then reproducible workflow questions is skipped
}
