% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_coding_prac}
\alias{w3_enforce_coding_prac}
\title{Enforce code frequency streaming on coding practices (wave 3)}
\usage{
w3_enforce_coding_prac(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then coding practices is skipped
}
