% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_degree_streaming}
\alias{w3_enforce_degree_streaming}
\title{Enforce degree streaming (wave 3)}
\usage{
w3_enforce_degree_streaming(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if highest qualification isn't degree then no degree subject has "Yes" returned
}
