% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_outside_role_streaming}
\alias{w3_enforce_outside_role_streaming}
\title{Enforce coding experience outside current role streaming}
\usage{
w3_enforce_outside_role_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if coding experience outside current role is no then follow up questions are skipped
}
