% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_prev_exp_streaming}
\alias{w3_enforce_prev_exp_streaming}
\title{Enforce previous coding experience streaming (wave 3)}
\usage{
w3_enforce_prev_exp_streaming(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if previous coding experience is no then where learned to code is skipped
}
