% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_rap_champ_knowl_streaming}
\alias{w3_enforce_rap_champ_knowl_streaming}
\title{Enforce heard of RAP streaming on RAP_champion and know_RAP_champ (wave 3)}
\usage{
w3_enforce_rap_champ_knowl_streaming(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if heard of RAP is "No" then are you RAP_champion and know_RAP_champ are skipped
}
