% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_rap_champ_streaming}
\alias{w3_enforce_rap_champ_streaming}
\title{Enforce RAP champion streaming (wave 3)}
\usage{
w3_enforce_rap_champ_streaming(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if RAP champion is yes then Do you know your RAP champion is skipped
}
