% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_3_preprocessing.R
\name{w3_enforce_streaming}
\alias{w3_enforce_streaming}
\title{API to enforce streaming rules (wave 3)}
\usage{
w3_enforce_streaming(data)
}
\arguments{
\item{data}{pre-processed data (wave 3)}
}
\value{
data
}
\description{
Enforce all streaming rules across all questions
}
