% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_4_preprocessing.R
\name{w4_enforce_skip_logic}
\alias{w4_enforce_skip_logic}
\title{enforce skip logic}
\usage{
w4_enforce_skip_logic(data, condition, skipped_cols)
}
\arguments{
\item{data}{data.frame}

\item{condition}{logical vector. Example: data$row == "skip response"}

\item{skipped_cols}{character. questions that should have been skipped if condition != TRUE}
}
\value{
data.frame with rows failing the check replaced with NAs
}
\description{
Replaces values in rows with NAs where check_skip_logic has identified backtracking.
}
