% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_4_preprocessing.R
\name{w4_enforce_streaming}
\alias{w4_enforce_streaming}
\title{Apply skip logic}
\usage{
w4_enforce_streaming(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Iteratively applies enforce_skip_logic to the necessary fields in the data.
}
