% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_5_preprocessing.R
\name{w5_clean_data}
\alias{w5_clean_data}
\title{Clean data}
\usage{
w5_clean_data(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
Recategorise department, workplace and first_learned data
}
