% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_5_preprocessing.R
\name{w5_clean_workplace}
\alias{w5_clean_workplace}
\title{Clean workplace data}
\usage{
w5_clean_workplace(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
reclassify 'other' text responses into CS/NHS
}
