% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_5_preprocessing.R
\name{w5_derive_language_status}
\alias{w5_derive_language_status}
\title{Derive language status}
\usage{
w5_derive_language_status(data)
}
\arguments{
\item{data}{tidied CARS wave 3 data (data.frame).}
}
\value{
data.frame
}
\description{
Derive the status of each programmming language as "access" (access only), "knowledge" (knowledge only), "both" or "neither".
}
