% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_5_preprocessing.R
\name{w5_derive_vars}
\alias{w5_derive_vars}
\title{Derive variables}
\usage{
w5_derive_vars(data)
}
\arguments{
\item{data}{tidied and relabelled CARS wave 3 dataset.}
}
\value{
data (data.frame).
}
\description{
API function for deriving additional variables.
}
