% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_5_preprocessing.R
\name{w5_rename_cols}
\alias{w5_rename_cols}
\title{Rename columns}
\usage{
w5_rename_cols(data)
}
\arguments{
\item{data}{tidy CARS dataset}
}
\value{
data.frame
}
\description{
Renames columns and removes unnecessary columns
}
