% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_apply_skip_logic}
\alias{w6_apply_skip_logic}
\title{Apply skip logic}
\usage{
w6_apply_skip_logic(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Applies skip logic to dataset, returning NA for downstream questions to adhere to survey streaming.
}
