% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_clean_data}
\alias{w6_clean_data}
\title{Clean data}
\usage{
w6_clean_data(data, config)
}
\arguments{
\item{data}{cleaned CARS dataset}

\item{config}{CARS config}
}
\value{
CARS dataset
}
\description{
Rename columns, enforce streaming
}
