% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_clean_departments}
\alias{w6_clean_departments}
\title{Clean department data}
\usage{
w6_clean_departments(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
add NHS to department list and merge departments where needed.
}
