% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_clean_first_learned}
\alias{w6_clean_first_learned}
\title{Clean first_learned data}
\usage{
w6_clean_first_learned(data)
}
\arguments{
\item{data}{cleaned CARS dataset}
}
\value{
CARS dataset
}
\description{
categorise 'other' responses.
}
