% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_derive_basic_rap_scores}
\alias{w6_derive_basic_rap_scores}
\title{Derive basic RAP scores}
\usage{
w6_derive_basic_rap_scores(data)
}
\arguments{
\item{data}{a date frame containing cleaned CARS wave 3 data}
}
\value{
dataframe containing the additional basic RAP score columns
}
\description{
Derive basic RAP score columns from existing variables and add to the dataframe.
}
