% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_derive_vars}
\alias{w6_derive_vars}
\title{Derive variables}
\usage{
w6_derive_vars(data)
}
\arguments{
\item{data}{tidied and relabelled CARS wave 3 dataset.}
}
\value{
data (data.frame).
}
\description{
function for deriving additional variables.
}
