% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_6_preprocessing.R
\name{w6_rename_cols}
\alias{w6_rename_cols}
\title{Rename columns}
\usage{
w6_rename_cols(data, config)
}
\arguments{
\item{data}{tidy CARS dataset}

\item{config}{CARS config}
}
\value{
data.frame
}
\description{
Renames columns and removes unnecessary columns
}
