% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-routing.R
\name{apply_skip_logic}
\alias{apply_skip_logic}
\title{Apply skip logic}
\usage{
apply_skip_logic(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Iteratively applies enforce_skip_logic to the necessary fields in the data.
}
