% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{calculate_multi_table_freqs}
\alias{calculate_multi_table_freqs}
\title{Create tidy cross table}
\usage{
calculate_multi_table_freqs(
  data,
  col1,
  col2,
  levels1,
  levels2,
  prop = TRUE,
  sample = FALSE
)
}
\arguments{
\item{data}{pre-processed CARS data set}

\item{col1}{first column of interest}

\item{col2}{column to cross-tabulate first column against}

\item{levels1}{factor levels for col1}

\item{levels2}{factor levels for col2}

\item{prop}{whether to return proportion data (0-1). TRUE by default. Assumes mutually exclusive response options.}

\item{sample}{returns proportion, count and, group size and sample size. FALSE by default}
}
\value{
data.frame
}
\description{
Returns a cross table in tidy data format.
}
