% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(data)
}
\arguments{
\item{data}{tidy CARS dataset}
}
\value{
clean CARS dataset
}
\description{
API function for data cleaning. Replaces column names with more suitable names and fixes incorrect apostrophes.
}
