% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(data, config)
}
\arguments{
\item{data}{cleaned CARS dataset}

\item{config}{CARS config}
}
\value{
CARS dataset
}
\description{
Rename columns, enforce streaming
}
