% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tables.R
\name{df_to_table}
\alias{df_to_table}
\title{Data frame to html table}
\usage{
df_to_table(data, config, question, crosstab = FALSE, column_headers)
}
\arguments{
\item{data}{list of CARS frequency tables, table names should match config}

\item{config}{CARS config}

\item{question}{string, question of interest. Name should match config.}

\item{crosstab}{whether to create a cross tabulation. FALSE by default}

\item{column_headers}{optional: list of column headers for the output table}
}
\value{
HTML table
}
\description{
converts dataframe to presentation HTML table
}
