% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tables.R
\name{df_to_table}
\alias{df_to_table}
\title{Data frame to html table}
\usage{
df_to_table(data, column_headers, crosstab = FALSE)
}
\arguments{
\item{data}{frequency table in tidy format. Last column should contain proportions.
First one or two columns should contain factors.}

\item{column_headers}{optional: list of column headers for the output table}

\item{crosstab}{whether to create a cross tabulation. FALSE by default}
}
\value{
HTML table
}
\description{
converts dataframe to presentation HTML table
}
