% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{get_all_waves}
\alias{get_all_waves}
\title{Get all CARS data, including previous waves}
\usage{
get_all_waves(mode = c("api", "file"))
}
\arguments{
\item{mode}{data source - "api"/"file"}
}
\value{
data frame
}
\description{
Ingest and preprocess all previous CARS data
}
