% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns}
\usage{
rename_cols(data)
}
\arguments{
\item{data}{tidy CARS dataset}
}
\value{
data.frame
}
\description{
Renames columns and removes unnecessary columns
}
