% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_cap_change_by_freq}
\alias{summarise_cap_change_by_freq}
\title{Summarise capability change by coding frequency}
\usage{
summarise_cap_change_by_freq(data)
}
\arguments{
\item{data}{full CARS datasetafter pre-processing}
}
\value{
frequency table (data.frame)
}
\description{
calculate the cross tab of coding frequency by capability change
}
