% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_cap_change_by_freq}
\alias{summarise_cap_change_by_freq}
\title{Summarise capability change by coding frequency}
\usage{
summarise_cap_change_by_freq(
  data,
  config,
  question1,
  question2,
  prop = TRUE,
  sample = FALSE
)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{returns proportion, count and, group size and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
calculate the cross tab of coding frequency by capability change
}
