% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_doc}
\alias{summarise_doc}
\title{Frequency of documentation use}
\usage{
summarise_doc(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of documentation use
}
