% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_rap_champ_status}
\alias{summarise_rap_champ_status}
\title{Knowledge of RAP Champions}
\usage{
summarise_rap_champ_status(data, sample = FALSE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
Create a frequency table of knowledge of RAP Champions
}
