% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-tables.R
\name{summarise_rap_opinions}
\alias{summarise_rap_opinions}
\title{Opinions of RAP}
\usage{
summarise_rap_opinions(data, config, question, prop = TRUE, sample = TRUE)
}
\arguments{
\item{data}{full CARS dataset after pre-processing}

\item{sample}{additionally returns count and sample size. FALSE by default}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of opinions of RAP
}
